Lms 116,118

+++ src/app/core/components/admin/users-page/user-out/user-out.component.ts
@@ -2,10 +2,12 @@ import {Component, Input, OnInit} from '@angular/core';
 import {MatDialog} from '@angular/material';
 
 import {IUser, IUserModel} from '../../interfaces';
-import {UserStatusEnum} from '../../../../../shared/enums';
+import {UserRolesEnum, UserStatusEnum} from '../../../../../shared/enums';
 import {CustomSnackbarService} from '../../../../../shared/services';
 import {AdminUsersService} from '../../services';
-import {UpdateUserComponent} from '../update-user/update-user.component';
+import {ChangeRoleUserComponent} from '../change-role-user/change-role-user.component';
+import {UpdateProfileComponent} from "../update-profile/update-profile.component";
+import {DeleteUserComponent} from "../delete-user/delete-user.component";
 
 @Component({
   selector: 'app-user-out',
@@ -15,6 +17,11 @@ import {UpdateUserComponent} from '../update-user/update-user.component';
 export class UserOutComponent implements OnInit {
   @Input() users: IUserModel;
   isBlocked = UserStatusEnum.BLOCKED;
+  roles = [
+    {name: 'Адміністратор', value: UserRolesEnum.ADMIN},
+    {name: 'Вчитель', value: UserRolesEnum.TEACHER},
+    {name: 'Студент', value: UserRolesEnum.STUDENT}
+  ];
 
   updatedAt: string = new Date().toString();
 
@@ -52,20 +59,43 @@ export class UserOutComponent implements OnInit {
   }
 
   delete(user: IUser): void {
-    const index: number = this.users.data.indexOf(user);
-    this.adminUsersService.deleteUser(user._id).subscribe(() => {
-      this.snackbarService.open('Користувач видалений');
-      this.users.data.splice(index, 1);
-    });
+    this.dialog.open(DeleteUserComponent, {
+      data: user
+    }).afterClosed().subscribe(value => {
+
+      if (value) {
+        const index: number = this.users.data.indexOf(user);
+        this.adminUsersService.deleteUser(user._id).subscribe(() => {
+          this.snackbarService.open('Користувач видалений');
+          this.users.data.splice(index, 1);
+        });
+      }
+    })
 
   }
 
   openEditForm(user: IUser) {
-    this.dialog.open(UpdateUserComponent, {
-      data: {
-        user,
-        users: this.users
-      }
+    this.dialog.open(ChangeRoleUserComponent, {
+      data: user
+    }).afterClosed().subscribe((role: UserRolesEnum) => {
+      const index: number = this.users.data.indexOf(user);
+      this.users.data[index].role_id = role;
     });
   }
+
+  updateProfile(user: IUser) {
+    this.dialog.open(UpdateProfileComponent, {
+      data: {user: user}
+    }).afterClosed().subscribe((value: IUser) => {
+      if (value) {
+        const index: number = this.users.data.indexOf(user);
+        this.users.data[index] = value;
+      }
+    })
+  }
+
+  private getNameOfRole(role: UserRolesEnum): string {
+    return this.roles[role - 1].name;
+  }
+
 }
